fix(wetland): patchtypes is real(r8), so convert before the patch_has_pft test
MOD_Const_LC declares the runtime patchtypes array as real(r8) even though it
is filled from an integer parameter table. Every existing use compares it
against a literal, which hides the type; passing it straight into
patch_has_pft(integer) does not.

Caught by gfortran, not by inspection:
  Error: Type mismatch in argument 'ptype'; passed REAL(8) to INTEGER(4)

+++ mkinidata/MOD_PercentagesPFTReadin.F90
@@ -43,7 +43,9 @@ CONTAINS
 #ifdef SinglePoint
       ! Wetland owns a WFT tile under WETLAND_PFT, so it must not bail out
       ! here: pftfrac would stay unset and the tile would carry no area.
-      IF (.not. patch_has_pft(patchtypes(SITE_landtype))) RETURN
+      ! nint: MOD_Const_LC declares the runtime patchtypes array as real(r8)
+      ! even though it is filled from an integer parameter table.
+      IF (.not. patch_has_pft(nint(patchtypes(SITE_landtype)))) RETURN
 #endif
 
 #ifndef SinglePoint
